fluss-server/src/test/java/com/alibaba/fluss/server/log/FetchParamsTest.java
/*
 * Copyright (c) 2024 Alibaba Group Holding Ltd.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.alibaba.fluss.server.log;

import com.alibaba.fluss.record.FileLogProjection;
import com.alibaba.fluss.record.TestData;

import org.junit.jupiter.api.Test;

import static org.assertj.core.api.AssertionsForClassTypes.assertThat;

/** Tests for {@link com.alibaba.fluss.server.log.FetchParams}. */
class FetchParamsTest {

    @Test
    void testSetCurrentFetch() {
        FetchParams fetchParams = new FetchParams(1, 100);
        fetchParams.setCurrentFetch(1L, 20L, 1024, TestData.DATA1_ROW_TYPE, null);
        assertThat(fetchParams.fetchOffset()).isEqualTo(20L);
        assertThat(fetchParams.maxFetchBytes()).isEqualTo(1024);
        assertThat(fetchParams.projection()).isNull();

        fetchParams.setCurrentFetch(2L, 30L, 512, TestData.DATA2_ROW_TYPE, new int[] {0, 2});
        assertThat(fetchParams.fetchOffset()).isEqualTo(30L);
        assertThat(fetchParams.maxFetchBytes()).isEqualTo(512);
        assertThat(fetchParams.projection()).isNotNull();

        FileLogProjection prevProjection = fetchParams.projection();

        fetchParams.setCurrentFetch(1L, 40L, 256, TestData.DATA1_ROW_TYPE, null);
        assertThat(fetchParams.projection()).isNull();

        fetchParams.setCurrentFetch(2L, 30L, 512, TestData.DATA2_ROW_TYPE, new int[] {0, 2});
        // the FileLogProjection should be cached
        assertThat(fetchParams.projection()).isNotNull().isSameAs(prevProjection);
    }
}
